% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.MMS.R
\name{calc.MMS}
\alias{calc.MMS}
\title{Calculate min-max scaling}
\usage{
calc.MMS(x, refval, reverse = FALSE)
}
\arguments{
\item{x}{Indicator values.}

\item{refval}{Refrence value.}

\item{reverse}{Reverse direction of indicator values.}
}
\description{
Calculate min-max scaling
}
