% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.Z.R
\name{calc.Z}
\alias{calc.Z}
\title{Calculate Z-score}
\usage{
calc.Z(x, reverse = FALSE)
}
\arguments{
\item{x}{Numeric variable.}

\item{reverse}{Should standardized variable be reversed?}
}
\description{
Calculate Z-score
}
