% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crodi2013.R
\docType{data}
\name{crodi2013}
\alias{crodi2013}
\title{Croatian development index 2013}
\format{A data frame with 556 rows and 9 variables:
\describe{
  \item{mb}{LAU-2 ID}
  \item{lau.name}{lau name}
  \item{county}{county name}
  \item{di}{development index}
  \item{inc}{income per capita}
  \item{rev}{lau budget per capita}
  \item{emp}{unemployment rate}
  \item{edu}{education rate}
  \item{pop}{population change}
}}
\usage{
crodi2013
}
\description{
Croatian development index 2013
}
\keyword{datasets}
