% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crodi2017.R
\docType{data}
\name{crodi2017}
\alias{crodi2017}
\title{Croatian development index 2017}
\format{A data frame with 556 rows and 10 variables:
\describe{
  \item{mb}{LAU-2 ID}
  \item{lau.name}{lau name}
  \item{county}{county name}
  \item{di}{development index}
  \item{inc}{income per capita}
  \item{rev}{lau budget per capita}
  \item{emp}{unemployment rate}
  \item{edu}{education rate}
  \item{pop}{population change}
  \item{agi}{index of aging}
}}
\usage{
crodi2017
}
\description{
Croatian development index 2017
}
\keyword{datasets}
