% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devindex.R
\name{devindex}
\alias{devindex}
\title{Calculate development index}
\usage{
devindex(formula, data, penalize = TRUE, wts = NULL, aggregate = "mean")
}
\arguments{
\item{formula}{Index variables}

\item{data}{Dataset}

\item{penalize}{Penalize composite index.}

\item{wts}{Indicator weights.}

\item{aggregate}{Method of aggregation of indicator values.}
}
\description{
Calculate development index
}
\details{
By default, function will calculate composite index penalized using Mazziotta-Pareto method. Setting \code{penalize = FALSE} will calculate composite index as a simple non-weighted mean of indicator values.

To use indicator weights specify \code{wts} as a named list.

Method of aggregation of indicator values is set via \code{aggregate} argument. Default is \code{mean}, but it can be a name of any function (typical alternative is \code{sum}).

By setting both \code{penalize = TRUE} and \code{wts} it is possible to use weighted indicators with penalization.
}
