% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize indicator}
\usage{
normalize(x, method = c("zscore", "mms"), reverse = FALSE, refval = NULL)
}
\arguments{
\item{x}{Vector of indicator values.}

\item{method}{Method of normalization.}

\item{reverse}{Reversing direction of indicator values.}

\item{refval}{If method is \code{mms}, a reference value used in normalization of indicator values.}
}
\description{
Normalize indicator
}
\details{
Two methods of normalization are available. Method \code{zscore} will standardize indicator values to z-scores with mean of 100 and standard deviation of 10. Method \code{mms} will use minimum-maximum normalization with respect to the lowest sample value, further divided by normalized value on the sample level.
}
\references{
Perisic & Wagner, Denona-Bogovic, Drezgic & Cegar.
}
