% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penalize.R
\name{penalize}
\alias{penalize}
\title{Calculate penalization}
\usage{
penalize(mf)
}
\arguments{
\item{mf}{Data frame with normalized indicator values}
}
\description{
Calculate penalization
}
\details{
Function will calculate penalty as defined in Mazziotta-Pareto method.
}
