% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refval2013.R
\docType{data}
\name{refval2013}
\alias{refval2013}
\title{Reference values for normalization (2013)}
\format{A list of length 5:
\describe{
  \item{inc}{income per capita}
  \item{rev}{lau budget per capita}
  \item{emp}{unemployment rate}
  \item{pop}{population change}
  \item{edu}{education rate}
}}
\usage{
refval2013
}
\description{
Reference values for normalization (2013)
}
\keyword{datasets}
